requirements.txt
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
bs4==0.0.1
cssselect==0.9.1
decorator==4.0.9
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
enum34==1.1.3
idna==2.1
ipaddress==1.0.16
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
MySQL-python==1.2.5
numpy==1.11.0
pathlib2==2.1.0
pbr==1.9.1
pexpect==4.0.1
pickleshare==0.7.2
psutil==3.4.2
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycparser==2.14
pygpgme==0.3
python-dateutil==2.5.3
pytz==2016.4
queuelib==1.4.2
requests==2.9.1
selenium==2.53.2
simplegeneric==0.8.1
six==1.10.0
stevedore==1.12.0
traitlets==4.2.1
Twisted==16.1.1
virtualenv-clone==0.2.6
w3lib==1.14.2
XlsxWriter==0.8.5
zope.interface==4.1.3
